% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addition-rules}
\alias{addition-rules}
\alias{+,spPlot,spTrack-method}
\alias{+,spTrack,spTrackGeom-method}
\title{Addition rules in spiralizePlus}
\usage{
\S4method{+}{spPlot,spTrack}(e1, e2)

\S4method{+}{spTrack,spTrackGeom}(e1, e2)
}
\arguments{
\item{e1}{A object defined in spiralizePlus}

\item{e2}{A object defined in spiralizePlus}
}
\value{
A object defined in spiralizePlus
}
\description{
spPlot(contain n spTracks)+spTrack=spPlot(contain n+1 spTracks),n>=0 \if{html}{\out{<br>}}
spTrak(contain n spTrakGeoms)+spTrackGeom=spTrack(contain n+1 spTrackGeoms),n>=0 \if{html}{\out{<br>}}
}
\examples{
NULL

}
