% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{show,spPlot-method}
\alias{show,spPlot-method}
\title{Draw the figures described by spPlot}
\usage{
\S4method{show}{spPlot}(object)
}
\arguments{
\item{object}{object of \linkS4class{spPlot}}
}
\value{
No return information
}
\description{
Draw the figures described by spPlot
}
\examples{
NULL
}
