% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprializePlus-package.R
\docType{package}
\name{spiralizePlus-package}
\alias{spiralizePlus}
\alias{spiralizePlus-package}
\title{spiralizePlus: ​Object-Oriented and Additive Operator-Based Spiralization}
\description{
spiralizePlus provides an object-oriented interface extension for the 'spiralize' package. It implements a modular architecture where spiral visualization components (axes, tracks, layers, etc.) are encapsulated as R4 objects with class-specific methods. Users construct complex spiral plots by combining these components through additive operators (+) in a declarative syntax.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TianzeLab/spiralizePlus}
  \item \url{https://tianzelab.github.io/spiralizePlus}
}

}
\author{
\strong{Maintainer}: Tianze Cao \email{hnrcao@qq.com} (\href{https://orcid.org/0000-0003-4308-1878}{ORCID})

Authors:
\itemize{
  \item Chen Liu \email{3169672818@qq.com}
}

}
\keyword{internal}
