% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\docType{class}
\name{spPlot-class}
\alias{spPlot-class}
\title{S4 class spPlot}
\description{
S4 class spPlot
}
\section{Slots}{

\describe{
\item{\code{initMode}}{character. It can only be the following values: "initialize", "initialize_by_gcoor", "initialize_by_time".}

\item{\code{initParams}}{list. A \strong{named} list that stores the parameters of the function called by the backend. Based on the value of initMode, the backend function will be one of the following four:\link[spiralize:spiral_initialize]{spiralize::spiral_initialize}, \link[spiralize:spiral_initialize_by_gcoor]{spiralize::spiral_initialize_by_gcoor}, \link[spiralize:spiral_initialize_by_time]{spiralize::spiral_initialize_by_time}.}

\item{\code{tracks}}{list. A list where \linkS4class{spTrack} are stored.}

\item{\code{clear}}{logical. Whether to call \link[spiralize:spiral_clear]{spiralize::spiral_clear} before drawing.}
}}

\examples{
NULL
}
