% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{spPlot}
\alias{spPlot}
\title{Object generator for S4 class spPlot}
\usage{
spPlot(initMode = "initialize", clear = TRUE, ...)
}
\arguments{
\item{initMode}{It can only be the following values:"initialize", "initialize_by_gcoor", "initialize_by_time".}

\item{clear}{Whether to call \link[spiralize:spiral_clear]{spiralize::spiral_clear} before drawing.}

\item{...}{
  Arguments passed on to \code{\link[spiralize:spiral_initialize]{spiralize::spiral_initialize}}, \code{\link[spiralize:spiral_initialize_by_gcoor]{spiralize::spiral_initialize_by_gcoor}}, \code{\link[spiralize:spiral_initialize_by_time]{spiralize::spiral_initialize_by_time}}
  \describe{
    \item{\code{xlim}}{Range on x-locations.}
    \item{\code{start}}{Start of the spiral, in degree. \code{start} and \code{end} should be positive and \code{start} should be smaller than \code{end}.}
    \item{\code{end}}{End of the spiral, in degree.}
    \item{\code{scale_by}}{How scales on x-axis are equally interpolated? The values can be one of "angle" and "curve_length". If
the value is "angle", equal angle difference corresponds to equal difference of data. In this case, in outer loops,
the scales are longer than in the inner loops, although the difference on the data are the same. If
the value is "curve_length", equal curve length difference corresponds to the equal difference of the data.}
    \item{\code{period}}{Under "angle" mode, the number of loops can also be controlled by argument \code{period} which controls the length
of data a spiral loop corresponds to. Note in this case, argument \code{end} is ignored and the value for \code{end} is
internally recalculated.}
    \item{\code{clockwise}}{Whether the curve is in a closewise direction. If it is set to \code{TRUE}, argument \code{flip} and \code{reverse} are ignored.}
    \item{\code{flip}}{How to flip the spiral? By default, the spiral starts from the origin of the coordinate and grows reverseclockwisely.
The argument controls the growing direction of the spiral.}
    \item{\code{reverse}}{By default, the most inside of the spiral corresponds to the lower boundary of x-location. Setting the value to \code{FALSE}
can reverse the direction.}
    \item{\code{polar_lines}}{Whether draw the polar guiding lines.}
    \item{\code{polar_lines_by}}{Increment of the polar lines. Measured in degree. The value can also be a vector that defines where to add polar lines.}
    \item{\code{polar_lines_gp}}{Graphics parameters for the polar lines.}
    \item{\code{padding}}{Padding of the plotting region. The value can be a \code{\link[grid:unit]{grid::unit()}} of length of one to two.}
    \item{\code{newpage}}{Whether to apply \code{\link[grid:grid.newpage]{grid::grid.newpage()}} before making the plot?}
    \item{\code{vp_param}}{A list of parameters sent to \code{\link[grid:viewport]{grid::viewport()}}.}
    \item{\code{unit_on_axis}}{Units on the axis.}
    \item{\code{normalize_year}}{Whether to enforce one spiral loop to represent a complete year?}
    \item{\code{period_per_loop}}{How many periods to put in a spiral loop?}
    \item{\code{verbose}}{Whether to print messages?}
  }}
}
\value{
Object \linkS4class{spPlot}
}
\description{
Object \linkS4class{spPlot} calls one of the following functions based on the value of initMode:\link[spiralize:spiral_initialize]{spiralize::spiral_initialize},\link[spiralize:spiral_initialize_by_gcoor]{spiralize::spiral_initialize_by_gcoor},\link[spiralize:spiral_initialize_by_time]{spiralize::spiral_initialize_by_time}.\if{html}{\out{<br>}}
The correct way to call it is as follows: \if{html}{\out{<br>}}
\code{spPlot(initMode = 'initialize',clear = TRUE,xlim = c(0,1),start = 360,end = 360 *5,scale_by = c("angle","curve_length"),period = NULL,clockwise = FALSE,flip = c("none","vertical","horizontal","both"),reverse = FALSE,polar_lines = scale_by=="angle",polar_lines_by = 30,polar_lines_gp = gpar(col = "#808080"),padding = unit(5,"mm"),newpage = TRUE,vp_param = list())} \if{html}{\out{<br>}} \if{html}{\out{<br>}}
\code{spPlot(initMode = 'initialize_by_gcoor',clear = TRUE,xlim = NULL,scal_by = "curve_length",...)} \if{html}{\out{<br>}} \if{html}{\out{<br>}}
\code{spPlot(initMode = 'initialize_by_time',clear = TRUE,xlim = NULL,start = NULL,end = NULL,unit_on_axis = c("days","months","weeks","hours","mins","secs"),period = c("years","months","weeks","days","hours","mins"),normalize_year = FALSE,period_per_loop = 1,polar_lines_by = NULL,verbose = TRUE,...)} \if{html}{\out{<br>}} \if{html}{\out{<br>}}
}
\examples{
NULL
}
