% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-geom.R
\name{spPoints}
\alias{spPoints}
\title{Add points on all sections of a single track.}
\usage{
spPoints(x = NULL, y = NULL, pch = 1, size = unit(0.4, "char"), gp = gpar())
}
\arguments{
\item{x}{X-locations of the data points.}

\item{y}{Y-locations of the data points.}

\item{pch}{Point type.}

\item{size}{Size of the points. Value should be a \code{\link[grid:unit]{grid::unit()}} object.}

\item{gp}{Graphical parameters.}
}
\value{
Object \linkS4class{spTrackGeom}
}
\description{
Object \linkS4class{spTrackGeom} will call the function \link[spiralize:spiral_points]{spiralize::spiral_points} while drawing.
}
\examples{
NULL
}
