% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\docType{class}
\name{spTrack-class}
\alias{spTrack-class}
\title{S4 class spTrack}
\description{
S4 class spTrack
}
\section{Slots}{

\describe{
\item{\code{func}}{character. Normally it is "spiral_track".}

\item{\code{params}}{list. A \strong{named} list that stores the parameters of the function called by the backend.}

\item{\code{trackGeoms}}{list. A list where \linkS4class{spTrackGeom} are stored.}
}}

\examples{
NULL
}
