% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{spTrack}
\alias{spTrack}
\title{Define a generic track}
\usage{
spTrack(
  ylim = c(0, 1),
  height = 0.8,
  background = TRUE,
  background_gp = gpar(fill = "#EEEEEE"),
  reverse_y = FALSE,
  gradient = FALSE
)
}
\arguments{
\item{ylim}{Data range of the y-locations.}

\item{height}{Height of the track. The value can be the fraction of the distance of the two neighbour spiral loops. The value can also be a \code{\link[grid:unit]{grid::unit()}} object.}

\item{background}{Whether to draw the background of the track, i.e. border and filled color of background.}

\item{background_gp}{Graphical parameters of the background.}

\item{reverse_y}{Whether reverse the direction of y-axis (i.e. pointing to the center of the spiral)?}

\item{gradient}{Whether draw the background in gradient? The value can be a positive integer of the number of gradients from \code{background_gp$fill} to white.}
}
\value{
Object \linkS4class{spTrack}
}
\description{
Object \linkS4class{spTrack} will call the function \link[spiralize:spiral_track]{spiralize::spiral_track} while drawing.
}
\examples{
NULL
}
